snapshot

+++ src/linear_algebra.rs
@@ -2,52 +2,52 @@ use std::ops;
 
 #[derive(Clone, Copy)]
 pub struct Matrix<const R: usize, const C: usize> {
-    pub inner: [[f32; C]; R]
+    pub inner: [[f64; C]; R]
 }
 
 pub type Vector2D = Matrix<2, 1>;
 
 impl Vector2D {
-    pub fn new_from_x_and_y(x: f32, y: f32) -> Self {
+    pub fn new_from_x_and_y(x: f64, y: f64) -> Self {
         Vector2D {
             inner: [[x], [y]],
         }
     }
 
     pub fn normalize(&self) -> Self {
-        let square_root = f32::sqrt(self.x()*self.x()+self.y()*self.y());
+        let square_root = f64::sqrt(self.x()*self.x()+self.y()*self.y());
         Vector2D::new_from_x_and_y(self.x()/square_root, self.y()/square_root)
     }
 
-    pub fn x(&self) -> f32 {
+    pub fn x(&self) -> f64 {
         self.inner[0][0]
     }
 
-    pub fn y(&self) -> f32 {
+    pub fn y(&self) -> f64 {
         self.inner[1][0]
     }
 }
 
-impl From<(f32, f32)> for Vector2D {
-    fn from(p: (f32, f32)) -> Self {
+impl From<(f64, f64)> for Vector2D {
+    fn from(p: (f64, f64)) -> Self {
         Self::new_from_x_and_y(p.0, p.1)
     }
 }
 
 impl<const R: usize, const C: usize> Matrix<R, C> {
-    pub fn new(inner: [[f32; C]; R]) -> Self {
+    pub fn new(inner: [[f64; C]; R]) -> Self {
         Matrix {inner}
     }
 
-    pub fn dot_product<const K: usize>(&self, row: usize, other: Matrix<C, K>, col: usize) -> f32 {
-        let mut sum: f32 = 0.;
+    pub fn dot_product<const K: usize>(&self, row: usize, other: Matrix<C, K>, col: usize) -> f64 {
+        let mut sum: f64 = 0.;
         for i in 0..C {
             sum += self.inner[row][i] * other.inner[i][col];
         }
         sum
     }
 
-    pub fn sum(&self) -> f32 {
+    pub fn sum(&self) -> f64 {
         let mut sum = 0.;
         for r in 0..R {
             for c in 0..C {
@@ -82,7 +82,7 @@ impl Matrix<2, 2> {
 impl<const R: usize, const C: usize, const K: usize> ops::Mul<Matrix<C, K>> for Matrix<R, C> {
     type Output = Matrix<R, K>;
     fn mul(self, rhs: Matrix<C, K>) -> Matrix<R, K> {
-        let mut inner: [[f32; K]; R] = [[0.; K]; R];
+        let mut inner: [[f64; K]; R] = [[0.; K]; R];
         for r in 0..R {
             for c in 0..K {
                 inner[r][c] = self.dot_product(r, rhs, c);
@@ -95,7 +95,7 @@ impl<const R: usize, const C: usize, const K: usize> ops::Mul<Matrix<C, K>> for 
 impl<const R: usize, const C: usize> ops::Add<Matrix<R, C>> for Matrix<R, C> {
     type Output = Matrix<R, C>;
     fn add(self, rhs: Matrix<R, C>) -> Matrix<R, C> {
-        let mut inner: [[f32; C]; R] = [[0.; C]; R];
+        let mut inner: [[f64; C]; R] = [[0.; C]; R];
         for r in 0..R {
             for c in 0..C {
                 inner[r][c] = self.inner[r][c] + rhs.inner[r][c];
@@ -112,10 +112,10 @@ impl<const R: usize, const C: usize> ops::Sub<Matrix<R, C>> for Matrix<R, C> {
     }
 }
 
-impl<const R: usize, const C: usize> ops::Mul<f32> for Matrix<R, C> {
+impl<const R: usize, const C: usize> ops::Mul<f64> for Matrix<R, C> {
     type Output = Matrix<R, C>;
-    fn mul(self, rhs: f32) -> Matrix<R, C> {
-        let mut inner: [[f32; C]; R] = [[0.; C]; R];
+    fn mul(self, rhs: f64) -> Matrix<R, C> {
+        let mut inner: [[f64; C]; R] = [[0.; C]; R];
         for r in 0..R {
             for c in 0..C {
                 inner[r][c] = self.inner[r][c] * rhs;
@@ -125,7 +125,7 @@ impl<const R: usize, const C: usize> ops::Mul<f32> for Matrix<R, C> {
     }
 }
 
-impl<const R: usize, const C: usize> ops::Mul<Matrix<R, C>> for f32 {
+impl<const R: usize, const C: usize> ops::Mul<Matrix<R, C>> for f64 {
     type Output = Matrix<R, C>;
     fn mul(self, rhs: Matrix<R, C>) -> Matrix<R, C> {
         rhs * self
